refactor move cloudfoundry to his own module
