+++ azure-spring-boot-bom/pom.xml
@@ -135,7 +135,17 @@
             </dependency>
             <dependency>
                 <groupId>com.microsoft.azure</groupId>
+                <artifactId>azure-cloudfoundry-spring-boot-starter</artifactId>
+                <version>${azure.spring.boot.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>com.microsoft.azure</groupId>
                 <artifactId>azure-spring-boot</artifactId>
+                <version>${azure.spring.boot.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>com.microsoft.azure</groupId>
+                <artifactId>azure-spring-boot-cloudfoundry</artifactId>
                 <version>${azure.spring.boot.version}</version>
             </dependency>
         </dependencies>
