+++ azure-spring-boot-cloudfoundry/src/main/java/com/microsoft/azure/cloudfoundry/VcapServiceConfig.java
@@ -3,7 +3,7 @@
  * Licensed under the MIT License. See LICENSE in the project root for
  * license information.
  */
-package com.microsoft.azure.spring.cloudfoundry.environment;
+package com.microsoft.azure.cloudfoundry;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
