+++ azure-spring-boot-cloudfoundry/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,1 @@
+org.springframework.boot.env.EnvironmentPostProcessor=com.microsoft.azure.cloudfoundry.VcapProcessor
