+++ azure-spring-boot-cloudfoundry/src/test/resources/vcap1.json
