+++ azure-spring-boot-cloudfoundry/src/test/resources/vcap2.json
