+++ azure-spring-boot-cloudfoundry/src/test/resources/vcap3.json
