+++ azure-spring-boot-samples/azure-cloud-foundry-service-sample/pom.xml
@@ -44,6 +44,10 @@
             <artifactId>spring-boot-starter-jersey</artifactId>
         </dependency>
         <dependency>
+            <groupId>com.microsoft.azure</groupId>
+            <artifactId>azure-cloudfoundry-spring-boot-starter</artifactId>
+        </dependency>
+        <dependency>
             <groupId>commons-io</groupId>
             <artifactId>commons-io</artifactId>
         </dependency>
