+++ azure-spring-boot-starters/azure-cloudfoundry-spring-boot-starter/README.md
@@ -1,0 +1,13 @@
+## Usage
+
+### Add the dependency
+
+`azure-cloudfoundry-spring-boot-starter` is published on Maven Central Repository.  
+If you are using Maven, add the following dependency.  
+
+```xml
+<dependency>
+    <groupId>com.microsoft.azure</groupId>
+    <artifactId>azure-cloudfoundry-spring-boot-starter</artifactId>
+</dependency>
+```
