+++ azure-spring-boot-starters/pom.xml
@@ -31,6 +31,7 @@
         <module>spring-data-gremlin-boot-starter</module>
         <module>azure-sqlserver-spring-boot-starter</module>
         <module>azure-servicebus-jms-spring-boot-starter</module>
+        <module>azure-cloudfoundry-spring-boot-starter</module>
     </modules>
 
 </project>
