+++ azure-spring-boot/src/main/resources/META-INF/spring.factories
@@ -1,4 +1,4 @@
-org.springframework.boot.env.EnvironmentPostProcessor=com.microsoft.azure.spring.cloudfoundry.environment.VcapProcessor,\
+org.springframework.boot.env.EnvironmentPostProcessor=\
 com.microsoft.azure.keyvault.spring.KeyVaultEnvironmentPostProcessor,\
 com.microsoft.azure.spring.autoconfigure.sqlserver.AlwaysEncryptedEnvironmentPostProcessor
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=com.microsoft.azure.spring.autoconfigure.cosmosdb.DocumentDBAutoConfiguration,\
