+++ pom.xml
@@ -51,6 +51,7 @@
         <module>azure-spring-boot</module>
         <module>azure-spring-boot-starters</module>
         <module>azure-spring-boot-samples</module>
+        <module>azure-spring-boot-cloudfoundry</module>
     </modules>
 
 </project>
